Refactor: Transition to TypeScript
This patch transitions UDO to TypeScript. It adds the tooling and
compiling support, as well as migrates all current React components
to TypeScript as well.

The `run` management command was removed, as well as all javascript.
Documentation was also updated.

Additionally, some features were introduced, primarily in an effort to
reduce dependencies.

  - The roster view now uses tabs and displays applicants more cleanly
    as a tab, along with each roster as a tab.
  - The Event view has been combined with the EventList view, so the
    event itself is seen in context with many other events.