Fix: Make the navbar more user focused
This patch adds some UI changes to make the logged in experience a
little clearer. In non-mobile mode, it makes the avatar and derby name
visible on all pages, with a drop down for other options. In mobile mode,
the hamburger menu shows the logged in user at the top of the menu.