Chore: Upgrade dependencies
This patch upgrades a number of dependencies. Python 3.7 brought in
a new plethora of changes and requirements. Some of the dependencies were
raising deprecation warnings in 3.7, for things that were going to go away
in Python 3.8. Some of those dependencies were upgraded.

When upgrading Django to 1.11, it raised a number of warnings were raised
for things that would be removed in Django 2.0. This patch addresses those
warnings, but it doesn't quite go far enough to suport Django 2.0; there
is more work to be done before that upgrade can be completed.

♬ : Chromeo / Sexy Socialite