Fix: Enforce maxLength on Profile
When editing a profile through the Profile view, it was possible to
cause backend errors by entering a string that was longer than the
database field (and the ORM) allowed.

While here, a toast message was added on succes of a saved profile, as
there was previously no successful feedback item.

Fixes #8