Fix: Unroster users when they withdraw
There was a bug where an official could be rostered, withdraw their
application, and they would still be rostered. It was made worse by the
fact that it's hidden from view on the staffing information because of
another bug (that might not be a bug now). This patch fixes that bug by
making sure that before an `ApplicationEntry` is deleted, the connected
official is unrostered from games.